move the assignment globally
if you need to access it everywhere in the file, this will allow for that. 
If you need something that can be returned by the function then you need to return an object or something that can
be accessed when the function executes i.e. ``` function(a, b){ does a thing... return { avatar: "whatever was assigned"} } ```